Adjust ParallelAttention to use Sliding Window Attention and Rolling Buffer Cache.
