--- megatron/model/mistral_model.py
@@ -1,622 +1,0 @@
-# coding=utf-8
-# Copyright (c) 2020, NVIDIA CORPORATION.  All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-Mistral Model
-Following implementation from huggingface, https://github.com/huggingface/transformers/blob/main/src/transformers/models/mistral/modeling_mistral.py
-"""
-
-import math
-from functools import partial
-
-import torch
-import torch.nn.functional as F
-
-from megatron import get_args
-from megatron import mpu
-from megatron.model.module import MegatronModule, float16_to_fp32, fp32_to_float16
-from megatron.model.enums import AttnMaskType, LayerType, AttnType
-from megatron.model.utils import get_linear_layer, init_method_normal, scaled_init_method_normal, attention_mask_func, \
-    openai_gelu, erf_gelu
-from megatron.model.fused_softmax import FusedScaleMaskSoftmax
-from megatron.model.language_model import Pooler
-from .cache import CacheView, RotatingBufferCache
-
-import deepspeed
-from deepspeed.accelerator import get_accelerator
-from deepspeed.pipe import PipelineModule, LayerSpec
-
-
-class RotaryEmbedding(torch.nn.Module):
-    def __init__(self, dim, max_position_embeddings=2048, base=10000, device=None):
-        super().__init__()
-
-        self.dim = dim
-        self.max_position_embeddings = max_position_embeddings
-        self.base = base
-        inv_freq = 1.0 / (self.base ** (torch.arange(0, self.dim, 2, dtype=torch.int64).float().to(device) / self.dim))
-        self.register_buffer("inv_freq", inv_freq, persistent=False)
-
-        # Build here to make `torch.jit.trace` work.
-        self._set_cos_sin_cache(
-            seq_len=max_position_embeddings, device=self.inv_freq.device, dtype=torch.get_default_dtype()
-        )
-
-    def _set_cos_sin_cache(self, seq_len, device, dtype):
-        self.max_seq_len_cached = seq_len
-        t = torch.arange(self.max_seq_len_cached, device=device, dtype=torch.int64).type_as(self.inv_freq)
-
-        freqs = torch.outer(t, self.inv_freq)
-        # Different from paper, but it uses a different permutation in order to obtain the same calculation
-        emb = torch.cat((freqs, freqs), dim=-1)
-        self.register_buffer("cos_cached", emb.cos().to(dtype), persistent=False)
-        self.register_buffer("sin_cached", emb.sin().to(dtype), persistent=False)
-
-    def forward(self, x, seq_len=None):
-        # x: [bs, num_attention_heads, seq_len, head_size]
-        if seq_len > self.max_seq_len_cached:
-            self._set_cos_sin_cache(seq_len=seq_len, device=x.device, dtype=x.dtype)
-
-        return (
-            self.cos_cached[:seq_len].to(dtype=x.dtype),
-            self.sin_cached[:seq_len].to(dtype=x.dtype),
-        )
-
-# Implement Sliding Window
-def modify_attention_mask_for_sliding_window(attention_mask, window_size):
-    """
-    Modifies the attention mask for sliding window attention.
-
-    Args:
-        attention_mask (torch.Tensor): Original attention mask of shape [batch_size, seq_len].
-        window_size (int): The size of the sliding window.
-
-    Returns:
-        torch.Tensor: Modified attention mask.
-    """
-    batch_size, seq_len = attention_mask.shape
-    new_mask = torch.full((batch_size, seq_len, seq_len), float('-inf'), device=attention_mask.device)
-
-    for i in range(seq_len):
-        start = max(0, i - window_size // 2)
-        end = min(seq_len, i + window_size // 2 + 1)
-        new_mask[:, i, start:end] = attention_mask[:, start:end]
-
-    return new_mask
-
-
-# Copied from transformers.models.llama.modeling_llama.rotate_half
-def rotate_half(x):
-    """Rotates half the hidden dims of the input."""
-    x1 = x[..., : x.shape[-1] // 2]
-    x2 = x[..., x.shape[-1] // 2 :]
-    return torch.cat((-x2, x1), dim=-1)
-
-def apply_rotary_pos_emb(q, k, cos, sin, position_ids, unsqueeze_dim=1):
-    """Applies Rotary Position Embedding to the query and key tensors.
-
-    Args:
-        q (`torch.Tensor`): The query tensor.
-        k (`torch.Tensor`): The key tensor.
-        cos (`torch.Tensor`): The cosine part of the rotary embedding.
-        sin (`torch.Tensor`): The sine part of the rotary embedding.
-        position_ids (`torch.Tensor`):
-            The position indices of the tokens corresponding to the query and key tensors. For example, this can be
-            used to pass offsetted position ids when working with a KV-cache.
-        unsqueeze_dim (`int`, *optional*, defaults to 1):
-            The 'unsqueeze_dim' argument specifies the dimension along which to unsqueeze cos[position_ids] and
-            sin[position_ids] so that they can be properly broadcasted to the dimensions of q and k. For example, note
-            that cos[position_ids] and sin[position_ids] have the shape [batch_size, seq_len, head_dim]. Then, if q and
-            k have the shape [batch_size, heads, seq_len, head_dim], then setting unsqueeze_dim=1 makes
-            cos[position_ids] and sin[position_ids] broadcastable to the shapes of q and k. Similarly, if q and k have
-            the shape [batch_size, seq_len, heads, head_dim], then set unsqueeze_dim=2.
-    Returns:
-        `tuple(torch.Tensor)` comprising of the query and key tensors rotated using the Rotary Position Embedding.
-    """
-    cos = cos[position_ids].unsqueeze(unsqueeze_dim)
-    sin = sin[position_ids].unsqueeze(unsqueeze_dim)
-    q_embed = (q * cos) + (rotate_half(q) * sin)
-    k_embed = (k * cos) + (rotate_half(k) * sin)
-    return q_embed, k_embed
-
-
-class MistralRMSNorm(torch.nn.Module):
-    def __init__(self, hidden_size, eps=1e-6):
-        """
-        MistralRMSNorm is equivalent to T5LayerNorm
-        """
-        super().__init__()
-        self.weight = torch.nn.Parameter(torch.ones(hidden_size))
-        self.variance_epsilon = eps
-
-    def forward(self, hidden_states):
-        input_dtype = hidden_states.dtype
-        hidden_states = hidden_states.to(torch.float32)
-        variance = hidden_states.pow(2).mean(-1, keepdim=True)
-        hidden_states = hidden_states * torch.rsqrt(variance + self.variance_epsilon)
-        return self.weight * hidden_states.to(input_dtype)
-
-# TODO adjust LlamaHead to Mistral if needed
-class MistralLMHead(MegatronModule):
-    """Causal LM head for Llama
-
-    Arguments:
-        vocab_size: size of vocabulary.
-        hidden_size: hidden size
-        gather_output: wether output logits being gathered or not.
-        init_method: init method for weight initialization
-    """
-
-    def __init__(self,
-                 hidden_size,
-                 vocab_size,
-                 init_method,
-                 parallel_output=True):
-        super(MistralLMHead, self).__init__()
-        args = get_args()
-        self.hidden_size = hidden_size
-        self.init_method = init_method
-        self.parallel_output = parallel_output
-
-        self.lm_head = mpu.ColumnParallelLinear(input_size=self.hidden_size,
-                                                output_size=vocab_size,
-                                                bias=False,
-                                                gather_output=not self.parallel_output,
-                                                skip_bias_add=True,
-                                                init_method=self.init_method, )
-
-    def forward(self, inputs):
-        logits, _ = self.lm_head(inputs)
-        return logits
-
-
-# TODO adjust LlamaHead to Mistral if needed
-class MistralLMHeadPipe(MistralLMHead):
-
-    def forward(self, inputs, **kwargs):
-        assert torch.is_tensor(inputs) or isinstance(inputs, tuple)
-        if isinstance(inputs, tuple):
-            hidden_states = inputs[0]
-        else:
-            hidden_states = inputs
-
-        if not hasattr(self, '_args'):
-            self._args = get_args()
-
-        if hasattr(self._args, 'attn_mask'):
-            attention_mask = None
-        else:
-            attention_mask = inputs[1]
-
-        logits = super().forward(hidden_states)
-
-        # If cmd args has attn_mask, we don't forward it as an activation.
-        if hasattr(self._args, 'attn_mask'):
-            return logits
-        else:
-            return logits, attention_mask
-
-
-class MistralEmbedding(MegatronModule):
-    """Language model embeddings.
-
-    Arguments:
-        hidden_size: hidden size
-        vocab_size: vocabulary size
-        init_method: weight initialization method
-    """
-
-    def __init__(self,
-                 hidden_size,
-                 vocab_size,
-                 init_method):
-        super(MistralEmbedding, self).__init__()
-        args = get_args()
-
-        self.hidden_size = hidden_size
-        self.init_method = init_method
-
-        # Word embeddings (parallel).
-        self.word_embeddings = mpu.VocabParallelEmbedding(vocab_size, self.hidden_size,
-                                                          init_method=self.init_method)
-
-    def forward(self, input_ids):
-        # Embeddings.
-        embeddings = self.word_embeddings(input_ids)
-        return embeddings
-
-class MistralEmbeddingPipe(MistralEmbedding):
-
-    def forward(self, inputs, **kwargs):
-        assert torch.is_tensor(inputs) or isinstance(inputs, tuple)
-        if isinstance(inputs, tuple):
-            input_ids = inputs[0]
-        else:
-            input_ids = inputs
-
-        if not hasattr(self, '_args'):
-            self._args = get_args()
-
-        if hasattr(self._args, 'attn_mask'):
-            attention_mask = None
-        else:
-            attention_mask = inputs[1]
-
-        embeddings = super().forward(input_ids)
-        # If cmd args has attn_mask, we don't forward it as an activation.
-        if hasattr(self._args, 'attn_mask'):
-            return embeddings
-        else:
-            return embeddings, attention_mask
-
-
-class MistralParallelMLP(MegatronModule):
-    """MLP.
-
-    MLP will take the input with h hidden state, project it to intermediate
-    hidden dimension, perform nonlinear transformation, and project the
-    state back into h hidden dimension.
-    """
-
-    def __init__(self, init_method, output_layer_init_method, moe=False, enable_expert_tensor_parallelism=False):
-        super(MistralParallelMLP, self).__init__()
-        args = get_args()
-        self.init_method = init_method
-        self.output_layer_init_method = output_layer_init_method
-
-        # Project to intermediate.
-        self.gate_proj = mpu.ColumnParallelLinear(
-            args.hidden_size,
-            args.ffn_hidden_size,
-            gather_output=False,
-            init_method=self.init_method,
-            skip_bias_add=True,
-            moe=moe,
-            enable_expert_tensor_parallelism=enable_expert_tensor_parallelism
-        )
-
-        self.up_proj = mpu.ColumnParallelLinear(
-            args.hidden_size,
-            args.ffn_hidden_size,
-            gather_output=False,
-            init_method=self.init_method,
-            skip_bias_add=True,
-            moe=moe,
-            enable_expert_tensor_parallelism=enable_expert_tensor_parallelism
-        )
-
-        self.activation_func = F.silu
-
-        # Project back to h.
-        self.down_proj = mpu.RowParallelLinear(
-            args.ffn_hidden_size,
-            args.hidden_size,
-            input_is_parallel=True,
-            init_method=self.output_layer_init_method,
-            skip_bias_add=True,
-            moe=moe,
-            enable_expert_tensor_parallelism=enable_expert_tensor_parallelism)
-
-    def forward(self, hidden_states):
-        intermediate_parallel = self.gate_proj(hidden_states)[0] * self.up_proj(hidden_states)[0]
-
-        intermediate_parallel = self.activation_func(intermediate_parallel)
-
-        output, _ = self.down_proj(intermediate_parallel)
-        return output
-
-
-# Copied from transformers.models.llama.modeling_llama.repeat_kv
-def repeat_kv(hidden_states: torch.Tensor, n_rep: int) -> torch.Tensor:
-    """
-    This is the equivalent of torch.repeat_interleave(x, dim=1, repeats=n_rep). The hidden states go from (batch,
-    num_key_value_heads, seqlen, head_dim) to (batch, num_attention_heads, seqlen, head_dim)
-    """
-    batch, num_key_value_heads, slen, head_dim = hidden_states.shape
-    if n_rep == 1:
-        return hidden_states
-    hidden_states = hidden_states[:, :, None, :, :].expand(batch, num_key_value_heads, n_rep, slen, head_dim)
-    return hidden_states.reshape(batch, num_key_value_heads * n_rep, slen, head_dim)
-
-# LlamaParallelAttention(MegatronModule)
-# LlamaParallelTransformerLayer(MegatronModule)
-# LlamaParallelTransformerLayerPipe(LlamaParallelTransformerLayer)
-# LlamaParallelTransformer(MegatronModule)
-
-class MistralParallelAttention(MegatronModule):
-    """Parallel self-attention layer abstract class.
-
-    Self-attention layer takes input with size [b, s, h]
-    and returns output of the same size.
-    """
-
-    def __init__(self, init_method,
-                 output_layer_init_method, layer_number,
-                 attention_type=AttnType.self_attn,
-                 attn_mask_type=AttnMaskType.causal,
-                 sliding_window_size=None,
-                 max_batch_size=None,
-                 n_kv_heads=None,
-                 n_layers=None,
-                 layer_id=None):
-        super(MistralParallelAttention, self).__init__()
-
-        assert attention_type == AttnType.self_attn
-        assert attn_mask_type == AttnMaskType.causal
-
-        self.sliding_window_size = sliding_window_size
-        self.n_kv_heads = n_kv_heads or self.num_attention_heads
-        self.head_dim = self.hidden_size_per_attention_head
-        self.max_batch_size = max_batch_size
-        self.n_layers = n_layers
-        self.layer_id = layer_id
-        self.cache = None
-
-        args = get_args()
-        self.fp16 = args.fp16
-        self.bf16 = args.bf16
-
-        self.apply_query_key_layer_scaling = args.apply_query_key_layer_scaling
-        self.attention_softmax_in_fp32 = args.attention_softmax_in_fp32
-        if self.apply_query_key_layer_scaling:
-            self.attention_softmax_in_fp32 = True
-        self.layer_number = max(1, layer_number)
-        self.attention_type = attention_type
-        self.attn_mask_type = attn_mask_type
-        self.init_method = init_method
-        self.output_layer_init_method = output_layer_init_method
-
-        self.num_attention_heads = args.num_attention_heads
-        projection_size = args.kv_channels * args.num_attention_heads
-
-        # Per attention head and per partition values.
-        world_size = mpu.get_tensor_model_parallel_world_size()
-        self.hidden_size_per_partition = mpu.divide(projection_size,
-                                                    world_size)
-        self.hidden_size_per_attention_head = mpu.divide(
-            projection_size, args.num_attention_heads)
-        self.num_attention_heads_per_partition = mpu.divide(
-            args.num_attention_heads, world_size)
-
-        # Strided linear layer.
-        if attention_type == AttnType.self_attn:
-            self.query_key_value = mpu.ColumnParallelLinear(
-                args.hidden_size,
-                3 * projection_size,
-                gather_output=False,
-                init_method=self.init_method)
-
-        coeff = None
-        self.norm_factor = math.sqrt(self.hidden_size_per_attention_head)
-        if self.apply_query_key_layer_scaling:
-            coeff = self.layer_number
-            self.norm_factor *= coeff
-
-        self.scale_mask_softmax = FusedScaleMaskSoftmax(
-            self.fp16, self.bf16,
-            self.attn_mask_type,
-            args.masked_softmax_fusion,
-            attention_mask_func,
-            self.attention_softmax_in_fp32,
-            coeff)
-
-        ## Rotary Position Embedding
-        self.rotary_emb = RotaryEmbedding(self.hidden_size_per_attention_head)
-
-        # Output.
-        self.dense = mpu.RowParallelLinear(
-            projection_size,
-            args.hidden_size,
-            input_is_parallel=True,
-            init_method=self.output_layer_init_method,
-            skip_bias_add=True)
-
-        if deepspeed.checkpointing.is_configured():
-            global get_cuda_rng_tracker, checkpoint
-            get_cuda_rng_tracker = deepspeed.checkpointing.get_cuda_rng_tracker
-            checkpoint = deepspeed.checkpointing.checkpoint
-
-    def forward(self, hidden_states, attention_mask, layer_past=None,
-                get_key_value=False):
-        # hidden_states: [sq, b, h]
-
-        ### Define RotatingBufferCache
-        if self.cache is None and self.sliding_window_size is not None:
-            self.cache = RotatingBufferCache(
-                n_layers=self.n_layers,
-                max_batch_size=self.max_batch_size,
-                sliding_window=self.sliding_window_size,
-                n_kv_heads=self.n_kv_heads,
-                head_dim=self.head_dim
-            ).to(hidden_states.device, hidden_states.dtype)
-
-        # =====================
-        # Query, Key, and Value
-        # =====================
-
-
-        if self.attention_type == AttnType.self_attn:
-            # Attention heads [sq, b, h] --> [sq, b, (np * 3 * hn)]
-            mixed_x_layer, _ = self.query_key_value(hidden_states)
-
-            # [sq, b, (np * 3 * hn)] --> [sq, b, np, 3 * hn]
-            new_tensor_shape = mixed_x_layer.size()[:-1] + \
-                               (self.num_attention_heads_per_partition,
-                                3 * self.hidden_size_per_attention_head)
-            mixed_x_layer = mixed_x_layer.view(*new_tensor_shape)
-
-            # [sq, b, np, 3 * hn] --> 3 [sq, b, np, hn]
-            (query_layer,
-             key_layer,
-             value_layer) = mpu.split_tensor_along_last_dim(mixed_x_layer, 3)
-
-        ### Sliding Window
-        if self.cache is not None and self.sliding_window_size is not None:
-            attention_mask = modify_attention_mask_for_sliding_window(
-                attention_mask, self.sliding_window_size
-            )
-
-        ### Cache
-        if self.cache is not None:
-            if layer_past is None:
-                self.cache.reset()
-                cache_metadata = self.cache.get_input_metadata(seqlens=[hidden_states.size(0)])
-                self.cache.init_kvseqlens(hidden_states.size(1))
-            else:
-                cache_metadata = self.cache.get_input_metadata(seqlens=[layer_past.size(0)])
-                self.cache.update_seqlens([hidden_states.size(0)])
-
-            cache_view = self.cache.get_view(layer_id=0, metadata=cache_metadata)
-            key_layer, value_layer = cache_view.interleave_kv(key_layer, value_layer)
-
-
-        # ==================================
-        # Rotary Position Embedding
-        # ==================================
-        # [sq, b, np, hn] --> [b, np, sq, hn] TODO optimize the permute of dimension back and forth
-        query_layer = query_layer.permute(1, 2, 0, 3)
-        key_layer = key_layer.permute(1, 2, 0, 3)
-        value_layer = value_layer.permute(1, 2, 0, 3)
-
-        cos, sin = self.rotary_emb(value_layer, seq_len=new_tensor_shape[0])
-        query_layer, key_layer = apply_rotary_pos_emb(query_layer, key_layer, cos, sin, offset=0)
-
-        # [b, np, sq, hn] --> [sq, b, np, hn] TODO optimize the permute of dimension back and forth
-        query_layer = query_layer.permute(2, 0, 1, 3).contiguous()
-        key_layer = key_layer.permute(2, 0, 1, 3).contiguous()
-        value_layer = value_layer.permute(2, 0, 1, 3).contiguous()
-
-        # ==================================
-        # Adjust key and value for inference
-        # ==================================
-
-        if layer_past is not None:
-            past_key, past_value = layer_past
-            key_layer = torch.cat((past_key.type_as(key_layer),
-                                   key_layer), dim=0)
-            value_layer = torch.cat((past_value.type_as(value_layer),
-                                     value_layer), dim=0)
-        if get_key_value:
-            present = (key_layer, value_layer)
-
-        # ===================================
-        # Raw attention scores. [b, np, s, s]
-        # ===================================
-
-        # [b, np, sq, sk]
-        output_size = (query_layer.size(1),
-                       query_layer.size(2),
-                       query_layer.size(0),
-                       key_layer.size(0))
-
-        # [sq, b, np, hn] -> [sq, b * np, hn]
-        query_layer = query_layer.view(output_size[2],
-                                       output_size[0] * output_size[1], -1)
-        # [sk, b, np, hn] -> [sk, b * np, hn]
-        key_layer = key_layer.view(output_size[3],
-                                   output_size[0] * output_size[1], -1)
-
-        # preallocting result tensor: [b * np, sq, sk]
-        matmul_result = torch.empty(
-            output_size[0] * output_size[1],
-            output_size[2],
-            output_size[3],
-            dtype=query_layer.dtype,
-            device=get_accelerator().current_device_name())
-
-        # Raw attention scores. [b * np, sq, sk]
-        matmul_result = torch.baddbmm(
-            matmul_result,
-            query_layer.transpose(0, 1),  # [b * np, sq, hn]
-            key_layer.transpose(0, 1).transpose(1, 2),  # [b * np, hn, sk]
-            beta=0.0, alpha=(1.0 / self.norm_factor))
-
-        # change view to [b, np, sq, sk]
-        attention_scores = matmul_result.view(*output_size)
-
-        # ==================================================
-        # Update attention mask for inference. [b, np, sq, sk]
-        # ==================================================
-
-        if get_key_value:
-            with torch.no_grad():
-                if layer_past is not None:
-                    attention_mask = attention_mask[
-                                     ...,
-                                     attention_scores.size(3) - 1,
-                                     :attention_scores.size(3)].unsqueeze(2)
-                else:
-                    attention_mask = attention_mask[
-                                     ...,
-                                     :attention_scores.size(3),
-                                     :attention_scores.size(3)]
-
-        # ===========================
-        # Attention probs and dropout
-        # ===========================
-
-        # attention scores and attention mask [b, np, sq, sk]
-        attention_probs = self.scale_mask_softmax(attention_scores,
-                                                  attention_mask)
-
-        # =========================
-        # Context layer. [sq, b, hp]
-        # =========================
-
-        # value_layer -> context layer.
-        # [sk, b, np, hn] --> [b, np, sq, hn]
-
-        # context layer shape: [b, np, sq, hn]
-        output_size = (value_layer.size(1),
-                       value_layer.size(2),
-                       query_layer.size(0),
-                       value_layer.size(3))
-
-        # change view [sk, b * np, hn]
-        value_layer = value_layer.view(value_layer.size(0),
-                                       output_size[0] * output_size[1], -1)
-
-        # change view [b * np, sq, sk]
-        attention_probs = attention_probs.view(output_size[0] * output_size[1],
-                                               output_size[2], -1)
-
-        # matmul: [b * np, sq, hn]
-        context_layer = torch.bmm(attention_probs, value_layer.transpose(0, 1))
-
-        # change view [b, np, sq, hn]
-        context_layer = context_layer.view(*output_size)
-
-        # [b, np, sq, hn] --> [sq, b, np, hn]
-        context_layer = context_layer.permute(2, 0, 1, 3).contiguous()
-
-        # [sq, b, np, hn] --> [sq, b, hp]
-        new_context_layer_shape = context_layer.size()[:-2] + \
-                                  (self.hidden_size_per_partition,)
-        context_layer = context_layer.view(*new_context_layer_shape)
-
-        # =================
-        # Output. [sq, b, h]
-        # =================
-
-        output, _ = self.dense(context_layer)
-
-        if get_key_value:
-            output = [output, present]
-
-        ### Trach cache for each layer
-        if self.cache is not None:
-            self.cache.get_view(self.layer_id).update(key_layer, value_layer)
-
-        return output
